Changed ingredient and item

+++ Restaurant POS/src/Model/Ingredient.java
@@ -5,12 +5,12 @@ public class Ingredient {
     
     public String name;
     public String description;
-    public int minimum;
-    public int quantity;
+    public double minimum;
+    public double quantity;
     public String unit;
     public String imgPath;
     
-    public Ingredient (String name, String description, int minimum, int quantity, String unit, String imgPath) {
+    public Ingredient (String name, String description, double minimum, double quantity, String unit, String imgPath) {
         this.name = name;
         this.description = description;
         this.minimum = minimum;
